refactor: break out generate file
